+++ lua/obsidian/pickers/_telescope.lua
@@ -2,6 +2,7 @@ local telescope = require "telescope.builtin"
 local telescope_actions = require "telescope.actions"
 local actions_state = require "telescope.actions.state"
 
+local api = require "obsidian.api"
 local Path = require "obsidian.path"
 local abc = require "obsidian.abc"
 local Picker = require "obsidian.pickers.picker"
@@ -13,6 +14,13 @@ local TelescopePicker = abc.new_class({
     return "TelescopePicker()"
   end,
 }, Picker)
+
+---@class obsidian.pickers.telescope_picker.CacheSelectedEntry
+---
+---@field value obsidian.cache.CacheNote[]
+---@field display string
+---@field ordinal string
+---@field absolute_path string
 
 ---@param prompt_bufnr integer
 ---@param keep_open boolean|?
@@ -124,6 +132,79 @@ local function attach_picker_mappings(map, opts)
   end
 end
 
+---Creates custom picker to search the notes using obsidian.cache.NoteCache
+---@param self obsidian.pickers.TelescopePicker
+---@param prompt_title string
+---@param opts obsidian.PickerFindOpts
+---@return obsidian.Picker
+local create_cache_picker = function(self, prompt_title, opts)
+  local pickers = require "telescope.pickers"
+  local finders = require "telescope.finders"
+  local config = require("telescope.config").values
+  local actions = require "telescope.actions"
+
+  local picker_opts = {
+    prompt_title = prompt_title,
+    attach_mappings = function(prompt_bufnr, map)
+      actions.select_default:replace(function()
+        local selection = get_entry(prompt_bufnr, false)
+
+        if not selection or not selection.absolute_path then
+          return
+        end
+
+        vim.schedule(function()
+          local open_cmd = api.get_open_strategy(Obsidian.opts.open_notes_in)
+          api.open_buffer(selection.absolute_path, { cmd = open_cmd })
+        end)
+      end)
+
+      attach_picker_mappings(map, {
+        entry_key = "absolute_path",
+        callback = opts.callback,
+        query_mappings = opts.query_mappings,
+        selection_mappings = opts.selection_mappings,
+      })
+      return true
+    end,
+  }
+
+  local cache_without_relative_path = vim.tbl_values(Obsidian.cache)
+  local workspace_path = Obsidian.dir.filename
+  return pickers.new(picker_opts, {
+    cwd = opts.dir,
+    finder = finders.new_table {
+      results = cache_without_relative_path,
+      ---@param entry obsidian.cache.CacheNote
+      ---@return obsidian.pickers.telescope_picker.CacheSelectedEntry
+      entry_maker = function(entry)
+        local concated_aliases = table.concat(entry.aliases, "|")
+        local concated_tags = table.concat(entry.tags, " #")
+        local relative_path = entry.absolute_path:gsub(workspace_path .. "/", "")
+        local display_name
+
+        if concated_aliases and concated_aliases ~= "" then
+          display_name = table.concat({ relative_path, concated_aliases }, "|")
+        else
+          display_name = relative_path
+        end
+
+        if Obsidian.opts.cache.show_tags and concated_tags and concated_tags ~= "" then
+          display_name = table.concat({ display_name, concated_tags }, " #")
+        end
+
+        return {
+          value = entry,
+          display = display_name,
+          ordinal = display_name,
+          absolute_path = entry.absolute_path,
+        }
+      end,
+    },
+    sorter = config.generic_sorter(),
+  })
+end
+
 ---@param opts obsidian.PickerFindOpts|? Options.
 TelescopePicker.find_files = function(self, opts)
   opts = opts or {}
@@ -134,24 +215,28 @@ TelescopePicker.find_files = function(self, opts)
     selection_mappings = opts.selection_mappings,
   }
 
-  telescope.find_files {
-    default_text = opts.query,
-    prompt_title = prompt_title,
-    cwd = opts.dir and tostring(opts.dir) or tostring(Obsidian.dir),
-    find_command = self:_build_find_cmd(),
-    attach_mappings = function(_, map)
-      attach_picker_mappings(map, {
-        callback = function(entry)
-          if opts.callback then
-            opts.callback(entry.filename)
-          end
-        end,
-        query_mappings = opts.query_mappings,
-        selection_mappings = opts.selection_mappings,
-      })
-      return true
-    end,
-  }
+  if opts.use_cache then
+    create_cache_picker(self, prompt_title, opts):find()
+  else
+    telescope.find_files {
+      default_text = opts.query,
+      prompt_title = prompt_title,
+      cwd = opts.dir and tostring(opts.dir) or tostring(Obsidian.dir),
+      find_command = self:_build_find_cmd(),
+      attach_mappings = function(_, map)
+        attach_picker_mappings(map, {
+          callback = function(entry)
+            if opts.callback then
+              opts.callback(entry.filename)
+            end
+          end,
+          query_mappings = opts.query_mappings,
+          selection_mappings = opts.selection_mappings,
+        })
+        return true
+      end,
+    }
+  end
 end
 
 ---@param opts obsidian.PickerGrepOpts|? Options.